build(qulice): upgrade to 0.30.3
Resolve all qulice violations surfaced by the version bump: rename
methods/variables to satisfy naming checks, remove unnecessary locals
and suppressions, migrate remaining JUnit4 tests to JUnit5, wrap
sockets in try-with-resources to satisfy PMD's CloseResource, reorder
class members to satisfy MethodsOrderCheck, and exclude qulice's own
error-prone scratch classes from the jacoco report so it no longer
collides with the real compiled classes.

Excludes the small set of UnitTestShouldIncludeAssert false positives
that qulice does not yet recognize (issues #1698, #1706).

Also bring the project into REUSE compliance (SPDX headers instead of
long-form license blocks, a REUSE.toml for non-code files) and fix
markdownlint violations in README.md, since the local pre-commit
hooks enforce both on every commit.

+++ src/main/java/org/cactoos/http/io/BoundedByteBuffer.java
@@ -1,31 +1,11 @@
 /*
- * The MIT License (MIT)
- *
- * Copyright (c) 2018 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
+ * SPDX-FileCopyrightText: Copyright (c) 2018-2026 Yegor Bugayenko
+ * SPDX-License-Identifier: MIT
  */
 package org.cactoos.http.io;
 
 /**
  * A very simple circular buffer of bytes.
- *
  * @since 0.1
  */
 public final class BoundedByteBuffer {
@@ -52,8 +32,7 @@ public final class BoundedByteBuffer {
 
     /**
      * Ctor.
-     *
-     * @param limit The size limit.
+     * @param limit The size limit
      */
     BoundedByteBuffer(final int limit) {
         this.internal = new byte[limit];
@@ -65,7 +44,6 @@ public final class BoundedByteBuffer {
     /**
      * Add a byte to the buffer, potentially by removing the oldest one to
      * satisfy the size limit.
-     *
      * @param add The byte to add
      */
     public void offer(final byte add) {
@@ -81,10 +59,9 @@ public final class BoundedByteBuffer {
 
     /**
      * Test if the buffer contains exactly the {@code bytes}.
-     *
-     * @param bytes The bytes to compare to.
+     * @param bytes The bytes to compare to
      * @return The value {@code true} if the buffer contains exactly
-     *  the {@code bytes}.
+     *  the {@code bytes}
      * @checkstyle ReturnCountCheck (14 lines)
      */
     @SuppressWarnings("PMD.OnlyOneReturn")
@@ -94,7 +71,7 @@ public final class BoundedByteBuffer {
         }
         int index = this.start;
         for (final byte current : bytes) {
-            if ((int) current != (int) this.internal[index]) {
+            if (current != this.internal[index]) {
                 return false;
             }
             index = (index + 1) % this.internal.length;
@@ -104,7 +81,7 @@ public final class BoundedByteBuffer {
 
     /**
      * Size of this buffer.
-     * @return This buffer's size.
+     * @return This buffer's size
      */
     private int size() {
         final int size;
@@ -120,7 +97,7 @@ public final class BoundedByteBuffer {
 
     /**
      * Is this buffer empty?
-     * @return True if empty, otherwise false.
+     * @return True if empty, otherwise false
      */
     private boolean empty() {
         return this.end == this.start && !this.full;
